### 2019-01-13 #1170 @whedon
[PRE REVIEW]: AxoPy: A Python Library for Implementing Human-Computer Interface Experiments
**Submitting author:** @ixjlyons (<a href="http://orcid.org/0000-0002-9143-8459">Kenneth Lyons</a>)
**Repository:** <a href="https://github.com/axopy/axopy" target ="_blank">https://github.com/axopy/axopy</a>
**Version:** v0.2.1
