### 2019-01-13 #1170 @arfon
@ixjlyons - looks like there is a problem compiling your paper. Please take a look and make your fixes in `master`. Whedon can then be asked to recompile the paper with `@whedon generate pdf`
