+++ paper/references.bib
@@ -10,7 +10,7 @@
 }
 
 @article{psychopy,
-    author = {Peirce, J. W.}
+    author = {Peirce, J. W.},
     title = {{PsychoPy - Psychophysics software in Python}},
     journal = {Journal of Neuroscience Methods},
     volume = {162},
@@ -36,8 +36,7 @@
 @inproceedings{OmearaEtAl-NER2019,
     author = {O'Meara, Sarah M. and Shyr, Megan C. and Lyons, Kenneth R. and
         Joshi, Sanjay S.},
-    title = {{Comparing Two Different Cursor Control
-            Methods which Use
+    title = {{Comparing Two Different Cursor Control Methods which Use
         Single-Site Surface Electromyography}},
     booktitle = {2019 9th International IEEE/EMBS Conference on Neural Engineering (NER)},
     address = {San Francisco, CA, USA},
