### 2019-01-13 ddca667
Fix up capitalization in paper references.
