+++ paper/references.bib
@@ -1,6 +1,7 @@
 @article{expyriment,
     author = {Krause, F. and Lindemann, O.},
-    title = {{Expyriment: A Python library for cognitive and neuroscientific experiments}},
+    title = {Expyriment: A {Python} library for cognitive and neuroscientific
+        experiments},
     journal = {Behavior Research Methods},
     volume = {46},
     number = {2},
@@ -11,7 +12,7 @@
 
 @article{psychopy,
     author = {Peirce, J. W.},
-    title = {{PsychoPy - Psychophysics software in Python}},
+    title = {{PsychoPy} - Psychophysics software in {Python}},
     journal = {Journal of Neuroscience Methods},
     volume = {162},
     number = {1--2},
@@ -22,10 +23,10 @@
 
 @inproceedings{LyonsJoshi-EMBC2018,
     author = {Lyons, Kenneth R. and Joshi, Sanjay S.},
-    title = {{Effects of Mapping Uncertainty on Visuomotor Adaptation to
-        Trial-By-Trial Perturbations with Proportional Myoelectric Control}},
-    booktitle = {2018 40th Annual International Conference of the IEEE
-        Engineering in Medicine and Biology Society (EMBC)},
+    title = {Effects of Mapping Uncertainty on Visuomotor Adaptation to
+        Trial-By-Trial Perturbations with Proportional Myoelectric Control},
+    booktitle = {2018 40th Annual International Conference of the {IEEE}
+        Engineering in Medicine and Biology Society ({EMBC})},
     pages = {5178-5181},
     month = {July},
     address = {Honolulu, HI, USA},
@@ -36,9 +37,10 @@
 @inproceedings{OmearaEtAl-NER2019,
     author = {O'Meara, Sarah M. and Shyr, Megan C. and Lyons, Kenneth R. and
         Joshi, Sanjay S.},
-    title = {{Comparing Two Different Cursor Control Methods which Use
-        Single-Site Surface Electromyography}},
-    booktitle = {2019 9th International IEEE/EMBS Conference on Neural Engineering (NER)},
+    title = {Comparing Two Different Cursor Control Methods which Use
+        Single-Site Surface Electromyography},
+    booktitle = {2019 9th International {IEEE/EMBS} Conference on Neural
+        Engineering ({NER})},
     address = {San Francisco, CA, USA},
     year = {2019}
 }
