### 2019-01-17 3511990
Clarified a few things in development docs.
