### 2019-01-17 #1170 @trallard
Hello @ixjlyons 👋 apologies for the delay in starting the review. Let's start by pinging a few fitting reviewers, with Python expertise as well as some background in signal processing, experimental design, or biomedical related data. We can start the review once two or three of them accept the request to review.

---

Dear potential reviewers (@lucask07, @ThomasA , @hooman650, @kczimm  ),

Based on your expertise I would like to ask if you can help us out with a submission to JOSS ([The Journal of Open Source Software](http://joss.theoj.org/)), where I’m an editor. The details of the submission are:

**Title**: AxoPy: A Python Library for Implementing Human-Computer Interface Experiments
**Summary:** Axo-Pythonic synapses are those in which an axon synapses upon a Python program. AxoPy aims to facilitate such connections between electrophysiolgical signals and machines by making it easy for researchers to develop human-computer interface experiments.
**Documentation**: [https://axopy.readthedocs.io/en/latest/index.html](https://axopy.readthedocs.io/en/latest/index.html)
**Article Proof**: [https://github.com/openjournals/joss-papers/blob/joss.01170/joss.01170/10.21105.joss.01170.pdf](https://github.com/openjournals/joss-papers/blob/joss.01170/joss.01170/10.21105.joss.01170.pdf)
**Submitting author:** @ixjlyons Kenneth Lyons
**Repository:** https://github.com/3dct/open_iA
**Version**: v0.2.1

@lucask07 : Lucas, are you able to and interested in taking on this review?
@ThomasA: Thomas, are you able to and interested in taking on this review?
@hooman650: Hooman, are you able to and interested in taking on this review?
@kczimm: Kendal, , are you able to and interested in taking on this review?

Your prompt response is appreciated. Please see [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information about reviewer criteria and the process if it is new to you.
