### 2019-01-19 #1170 @ixjlyons
@trallard: I noticed that @peircej is in the spreadsheet of reviewers --- he's the author of PsychoPy, which is a fairly closely related experiment-building framework. I imagine he'd be a good reviewer for AxoPy if he has the time.

Also, the link to the repository should be: https://github.com/axopy/axopy.
