### 2019-01-19 #1170 @trallard
thanks @lucask07

@peterjc would you be able and interested in reviewing this submission for JOSS?
