### 2019-01-20 #1170 @peircej
I'd be happy to review if still needed
