### 2019-01-21 #1191 @whedon
[REVIEW]: AxoPy: A Python Library for Implementing Human-Computer Interface Experiments
**Submitting author:** @ixjlyons (<a href="http://orcid.org/0000-0002-9143-8459">Kenneth Lyons</a>)
**Repository:** <a href="https://github.com/axopy/axopy" target ="_blank">https://github.com/axopy/axopy</a>
**Version:** 0.2.3
