### 2019-01-21 #1191 @trallard
Some important notes about the revision:
- We ideally want the review to be conducted within a couple of weeks, but if any of you are struggling for time, please let me know
- As you go over the checklist: if you encounter any issues or suggest any changes to the package feel free to use this issue as a discussion platform as well as opening issues on the submission repo as needed.
- If there is anything that is not clear in the reviewers manual or you have any questions about the review process, please feel free to ping me here
