### 2019-01-26 #1191 @lucask07
Hi @ixjlyons, this package is looking good, very nice work.

Installation via pip (OSX Python 3.7) proceeded without issues:
```terminal
python -m pip install axopy
```
However, when importing ``axopy`` I hit a snag.

```python
import axopy
ImportError: cannot import name 'Transmitter' from 'axopy.messaging'
```

Note that this works:
```python
from axopy.messaging import transmitter
```
so I'm suspecting just a capatalization issue?
