### 2019-01-26 #1191 @lucask07
Thanks @ixjlyons
The import of *axopy* is working now.

I'm attempting to run the example within the README file. The QT window seems stuck and only reports "Ready". I tried this in an iPython window and without iPython. It would be helpful to know the expected output of this example. I'm running MacOSX.

A screenshot is below:
![screen shot 2019-01-26 at 12 40 08 pm](https://user-images.githubusercontent.com/1977199/51791696-315eed00-216c-11e9-883b-cd589b79d957.png)
