### 2019-01-26 #1191 @trallard
Hey @ixjlyons we can definitely bump the version in JOSS. I'll wait until the review is completed in case the reviewers request any more substantial changes or you make another release in the meantime
