### 2019-01-26 axopy/axopy#55 @ixjlyons
Application doesn't exit nicely when running in an interactive shell
Running the following example line-by-line from the README in an interactive shell results in the experiment window remaining open despite control going back to the shell:

```python
import axopy

daq = axopy.daq.NoiseGenerator(rate=1000, num_channels=4, read_size=100)
exp = axopy.experiment.Experiment(daq=daq)
exp.run(axopy.task.Oscilloscope())
```

The window stays open and can't be closed (at least on my machine) without killing the python process that called `exp.run(...)`.

I believe this is fixable and it's probably worthwhile to look into it because users of Spyder or new users running the example in an interactive session would have this problem.
