### 2019-01-26 #1191 @ixjlyons
@lucask07 yes, that looks like it's working as expected. You should be able to press `Enter` (with the Qt window focused) and then the screen should show an "oscilloscope" with some random signals updating periodically. Pressing `Enter` again should then finish the experiment and close the window. One thing to note is running an experiment in an interactive shell isn't something I've intended to support, so it looks like there are a few buggy behaviors when doing so. I foresee this as potentially being an issue for people using an IDE that runs things in an interactive shell like Spyder, so I'll work on finding a solution and track things [here](https://github.com/axopy/axopy/issues/55).
