### 2019-01-26 #1191 @ixjlyons
Ah, that appears to be an upstream issue. There is a way to selectively disable that operation, but I'm currently leaning toward disabling the context menu altogether. It adds a few nice features, but it's probably best if AxoPy users implement their own processing in the form of the `pipeline` keyword arg in creating the `Oscilloscope` task.
