### 2019-01-27 #1191 @lucask07
Hi @ixjlyons,
Thank you for this software it looks great. The documentation is very nicely done. 😀
My only suggestion would be to emphasize examples that involve connecting a real-world DAQ.
The [pymcc](https://github.com/axopy/pymcc/blob/master/pymcc/mccdaq.py) is a useful example, but explicitly knowing things like the needed format of the return data from the `read` method. I'm thinking about using the `Oscilloscope` connected to a USB3.0 DAQ system that transfers ~20 MB/s.

So two questions:

1. What format must the data returned by `read` be?
2. What is the maximum data rate that the `oscilloscope` waveform display can keep up with?
