### 2019-01-27 #1191 @peircej
I agree. The paper looks all fine. The software looks sensible, useful and following best practices.
Minor suggestions around docs:
- I also hit the snag of not realising that I should hit <return> to start simulating the data collection. (I opened the script file and found it mentioned inside the script, but I did this *after* trying to run). Rather than adding this in the documentation, I would add it to the actual screen "Ready/nHit <return> to start". Lots of people will jump into a demo before reading anything.
- I'd also include the demos in the package itself (the first place I looked for demos was site-packages and then went online and had to download them separately)
- Lastly, for projects where the documentation isn't directly at the home page (e.g. hope page is github) it's best to include a link to the readthedocs documentation very clearly at the top of the readme file. When someone arrives at https://github.com/axopy/axopy what steps will they be taking to find the (formatted) docs?

But these are minor things. Useful package, put together nicely (and I love the logo!)

I also accept this submission.
