### 2019-01-28 80726d0
Disable the context menu in Oscilloscope.
I actually figured out you could disable just the FFT menu item
selectively but I think it's probably best to just disable it. Might be
an interesting idea to facilitate some way to add your own menu to
enable/disable pipelines, for example.
