### 2019-01-28 f1b6494
Say what to do on the ready screen.
