### 2019-01-29 7196bf2
Add PyPI and conda-forge badges. Make link to docs prominent.
