### 2019-01-29 d1518ac
Add sphinx-gallery extension to render examples in docs.
