+++ examples/README.txt
@@ -1,0 +1,6 @@
+========
+Examples
+========
+
+Here are some complete examples that can be run. Click on an example to see the
+source code and some comments about the example.
