+++ docs/index.rst
@@ -20,5 +20,6 @@ Documentation Contents
 
    installation
    user_guide
+   auto_examples/index
    api
    development
