### 2019-01-29 d30fb5a
Include a bit more info in DAQ API.
