+++ docs/user_guide/daq.rst
@@ -67,12 +67,17 @@ API as well. This is defined below:
 
 .. code-block:: text
 
-   Daq
-      start - Called once before the first read.
-      read  - Request a new buffer of data from the hardware. Parameters (like
-              the size of the buffer or number of samples to read should be
-              set up in the daq constructor.
-      stop  - Called when the user wants the device to stop reading data.
+   DAQ API
+      start() - Called once before the first read.
+
+      read()  - Request a new buffer of data from the hardware. Parameters (like
+                the size of the buffer or number of samples to read) should be
+                set up in the daq constructor. Canonical AxoPy devices generate
+                a NumPy ndarray with shape (n_channels, n_samples), but the only
+                real restriction is that your Pipelines can consume the data
+                generated by your device.
+
+      stop()  - Called when the user wants the device to stop reading data.
 
 In addition, the DAQ implementation should raise an ``IOError`` if something
 goes wrong during data acquisition.
