### 2019-01-29 #1191 @ixjlyons
Thank you both @peircej and @lucask07 for your reviews. In a nutshell, I agree with all proposed changes and have made them. Here they are as I've understood:

- There was a packaging error that caused v0.2.1 to be broken.
- A context menu provided by an upstream library caused a crash. I disabled the menu, at least until the upstream issue is fixed.
- It was totally unclear what to do upon running the example in the README. I added instructions to the "ready" screen telling the user what to do to proceed. Thanks @peircej for the much more elegant method of conveying that info :)
- The DAQ API didn't say anything about the expected format of the data. I added some info about this in the documentation.
- Links to documentation were fairly hidden in the README. I added a link at the top.
- The examples were not very easy to find/obtain, given that they weren't packaged and weren't mentioned in the documentation. I can see some benefits to including the examples in the package itself (so it can be run like `python -m axopy.examples.example`), but for now I've opted to use sphinx-gallery to render the examples in the documentation.

The changes can be seen here: https://github.com/axopy/axopy/compare/181eab2d2a9f203c61b0a82b259538cdf838d890...master

----

@lucask07 to answer your questions:

> I'm thinking about using the Oscilloscope connected to a USB3.0 DAQ system that transfers ~20 MB/s.
>
> So two questions:
>
>    What format must the data returned by read be?
>    What is the maximum data rate that the oscilloscope waveform display can keep up with?

I added a small note about this in the docs, but basically it's up to you. So far in AxoPy, there's an "encouraged" return type: ndarray with shape (n_channels, n_samples). I say encouraged because you could always just implement some transformations as a pipeline block to get it into the format needed by Oscilloscope :). This may become a problem, however, if that can't be done fast enough with the addition of plotting time... AxoPy uses [pyqtgraph](http://www.pyqtgraph.org/) in the Oscilloscope -- you'd likely have to downsample, but it can plot quite fast (there are some pyqtgraph examples that demonstrate). The process of creating the ndarray may be the problem at that rate. If you try it, let me know.
