### 2019-02-11 #1191 @peircej
Minor: I think the request above was meant for the author @ixjlyons rather than me (to update the version etc).
