### 2019-02-12 #1191 @ixjlyons
Ok, here's the DOI for axopy-0.2.3: [10.5281/zenodo.2562630](https://doi.org/10.5281/zenodo.2562630)

Thanks once again @trallard, @peircej, and @lucask07 for your time and feedback.
