### 2019-02-16 #1191 @trallard
@ixjlyons all looks good to me so I am going to proceed for acceptance

@lucask07 and @peircej thanks a lot for your time and valuable contribution to JOSS as reviewers for this submission 🙌🏻

@arfon: this submission is accepted and ready to be published 🎉👾
