### 2019-02-16 #1191 @arfon
> @arfon: this submission is accepted and ready to be published 🎉👾

@trallard - in the future could you please ping @openjournals/joss-eics when a review is complete? That way, the managing editor on duty can process the submission.
